+++ packages/react/src/primitives/Select/Select.tsx
@@ -4,10 +4,11 @@ import classNames from 'classnames';
 import { classNameModifier, classNameModifierByFlag } from '../shared/utils';
 import { ComponentClassNames } from '../shared/constants';
 import { Flex } from '../Flex';
-import { IconExpandMore } from '../Icon/internal';
+import { IconExpandMore, useIcons } from '../Icon';
 import { ForwardRefPrimitive, Primitive } from '../types';
 import { BaseSelectProps, SelectProps } from '../types/select';
 import { View } from '../View';
+import { useFieldset } from '../Fieldset/useFieldset';
 
 const SelectPrimitive: Primitive<SelectProps, 'select'> = (
   {
@@ -18,7 +19,7 @@ const SelectPrimitive: Primitive<SelectProps, 'select'> = (
     value,
     defaultValue,
     hasError,
-    icon = <IconExpandMore />,
+    icon,
     iconColor,
     children,
     placeholder,
@@ -40,6 +41,8 @@ const SelectPrimitive: Primitive<SelectProps, 'select'> = (
     classNameModifierByFlag(ComponentClassNames.Select, 'error', hasError),
     className
   );
+  const icons = useIcons('select');
+  const { isFieldsetDisabled } = useFieldset();
 
   return (
     <View className={ComponentClassNames.SelectWrapper}>
@@ -53,7 +56,7 @@ const SelectPrimitive: Primitive<SelectProps, 'select'> = (
             ? DEFAULT_PLACEHOLDER_VALUE
             : defaultValue
         }
-        isDisabled={isDisabled}
+        isDisabled={isFieldsetDisabled ? isFieldsetDisabled : isDisabled}
         required={isRequired}
         data-size={size}
         data-variation={variation}
@@ -71,7 +74,7 @@ const SelectPrimitive: Primitive<SelectProps, 'select'> = (
         )}
         color={iconColor}
       >
-        {icon}
+        {icon ?? icons?.expand ?? <IconExpandMore />}
       </Flex>
     </View>
   );