+++ packages/react/src/primitives/Select/__tests__/Select.test.tsx
@@ -3,6 +3,7 @@ import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 
 import { Select } from '../Select';
+import { Fieldset } from '../../Fieldset';
 import { IconExpandMore } from '../../Icon/internal';
 import { ComponentClassNames } from '../../shared';
 
@@ -123,6 +124,28 @@ describe('Select primitive test suite', () => {
     const select = await screen.findByTestId('test-select');
     expect(select).toBeDisabled();
     expect(select).toBeRequired();
+  });
+
+  it('should always be disabled if parent Fieldset isDisabled', async () => {
+    render(
+      <Fieldset legend="legend" isDisabled>
+        <Select testId="select">
+          <option value="1">1</option>
+          <option value="2">2</option>
+          <option value="3">3</option>
+        </Select>
+        <Select testId="selectWithDisabledProp" isDisabled={false}>
+          <option value="1">1</option>
+          <option value="2">2</option>
+          <option value="3">3</option>
+        </Select>
+      </Fieldset>
+    );
+
+    const select = await screen.findByTestId('select');
+    const selectDisabled = await screen.findByTestId('selectWithDisabledProp');
+    expect(select).toHaveAttribute('disabled');
+    expect(selectDisabled).toHaveAttribute('disabled');
   });
 
   it('should render placeholder correctly if it is set', async () => {